+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -257,6 +257,12 @@ New checks
 
   Finds cyclical initialization of static variables.
 
+- New :doc:`modernize-use-if-consteval
+  <clang-tidy/checks/modernize/use-if-consteval>` check.
+
+  Replaces direct ``std::is_constant_evaluated()`` checks in ``if`` statements
+  with C++23's ``if consteval`` syntax.
+
 - New :doc:`modernize-use-std-bit
   <clang-tidy/checks/modernize/use-std-bit>` check.